refactor(middleware): move ResponseCacheMiddleware into _internal
